apollo_bot_without_API/requirements.txt
requests
gspread
oauth2client
python-dotenv
Flask
